Change input buffer type into uint8_t

+++ color.hpp
@@ -2,5 +2,5 @@
 #include <cstdint>
 #include <vector>
 
-void rgb2ycbcr(int16_t *in, int width);
-void subsample(int16_t *in, std::vector<int16_t *> out, int width, int YCCtype);
+void rgb2ycbcr(uint8_t *in, int width);
+void subsample(uint8_t *in, std::vector<int16_t *> out, int width, int YCCtype);
